Move helper class to a separate package and emphasize that it is experimental.
